--- src/components/blog/Breadcrumbs/index.ts
@@ -1,3 +1,0 @@
-import Breadcrumbs from "./Breadcrumbs";
-
-export default Breadcrumbs;
