+++ src/components/blog/ExplorerSection/OpenPostList/OpenPostListItem/index.style.ts
