--- src/components/blog/ExplorerSection/OpenPostList/OpenPostListItem/index.ts
@@ -1,3 +1,0 @@
-import OpenPostListItem from "./OpenPostListItem";
-
-export default OpenPostListItem;
