--- src/components/blog/ExplorerSection/OpenPostList/index.ts
@@ -1,3 +1,0 @@
-import OpenPostList from "./OpenPostList";
-
-export default OpenPostList;
