--- src/components/blog/ExplorerSection/index.ts
@@ -1,3 +1,0 @@
-import ExplorerSection from "./ExplorerSection";
-
-export default ExplorerSection;
