+++ src/components/blog/OpenPostTabs/OpenPostTab/index.style.ts
