--- src/components/blog/OpenPostTabs/OpenPostTab/index.ts
@@ -1,3 +1,0 @@
-import OpenPostTab from "./OpenPostTab";
-
-export default OpenPostTab;
