+++ src/components/blog/OpenPostTabs/index.style.ts
