--- src/components/blog/OpenPostTabs/index.ts
@@ -1,3 +1,0 @@
-import OpenPostTabs from "./OpenPostTabs";
-
-export default OpenPostTabs;
