--- src/components/blog/PostTemplate/index.ts
@@ -1,3 +1,0 @@
-import PostTemplate from "./PostTemplate";
-
-export default PostTemplate;
